GSON and JDA event collisions resolved, PR review changes
